requirements.txt
boto3
Flask
requests
python-dotenv
